+++ lib/internal/bootstrap/node.js
@@ -44,33 +44,57 @@
 
     setupGlobalVariables();
 
-    const _process = NativeModule.require('internal/process');
-    _process.setupConfig(NativeModule._source);
-    _process.setupSignalHandlers();
-    _process.setupUncaughtExceptionCapture(exceptionHandlerState,
-                                           _shouldAbortOnUncaughtToggle);
+    // Bootstrappers for all threads, including worker threads and main thread
+    const perThreadSetup = NativeModule.require('internal/process/per_thread');
+    // Bootstrappers for the main thread only
+    let mainThreadSetup;
+    // Bootstrappers for the worker threads only
+    let workerThreadSetup;
+    if (isMainThread) {
+      mainThreadSetup = NativeModule.require(
+        'internal/process/main_thread_only'
+      );
+    } else {
+      workerThreadSetup = NativeModule.require(
+        'internal/process/worker_thread_only'
+      );
+    }
+
+    perThreadSetup.setupAssert();
+    perThreadSetup.setupConfig(NativeModule._source);
+
+    if (isMainThread) {
+      mainThreadSetup.setupSignalHandlers();
+    }
+
+    perThreadSetup.setupUncaughtExceptionCapture(exceptionHandlerState,
+                                                 _shouldAbortOnUncaughtToggle);
+
     NativeModule.require('internal/process/warning').setup();
     NativeModule.require('internal/process/next_tick').setup(_setupNextTick,
                                                              _setupPromises);
-    NativeModule.require('internal/process/stdio').setup();
-    NativeModule.require('internal/process/methods').setup(_chdir,
-                                                           _umask,
-                                                           _initgroups,
-                                                           _setegid,
-                                                           _seteuid,
-                                                           _setgid,
-                                                           _setuid,
-                                                           _setgroups);
+
+    if (isMainThread) {
+      mainThreadSetup.setupStdio();
+      mainThreadSetup.setupProcessMethods(
+        _chdir, _umask, _initgroups, _setegid, _seteuid,
+        _setgid, _setuid, _setgroups
+      );
+    } else {
+      workerThreadSetup.setupStdio();
+    }
 
     const perf = process.binding('performance');
     const {
       NODE_PERFORMANCE_MILESTONE_BOOTSTRAP_COMPLETE,
     } = perf.constants;
 
-    _process.setup_hrtime(_hrtime, _hrtimeBigInt);
-    _process.setup_cpuUsage(_cpuUsage);
-    _process.setupMemoryUsage(_memoryUsage);
-    _process.setupKillAndExit();
+    perThreadSetup.setupRawDebug(_rawDebug);
+    perThreadSetup.setupHrtime(_hrtime, _hrtimeBigInt);
+    perThreadSetup.setupCpuUsage(_cpuUsage);
+    perThreadSetup.setupMemoryUsage(_memoryUsage);
+    perThreadSetup.setupKillAndExit();
+
     if (global.__coverage__)
       NativeModule.require('internal/process/write-coverage').setup();
 
@@ -90,10 +114,9 @@
       NativeModule.require('internal/inspector_async_hook').setup();
     }
 
-    if (isMainThread)
-      _process.setupChannel();
-
-    _process.setupRawDebug(_rawDebug);
+    if (isMainThread) {
+      mainThreadSetup.setupChildProcessIpcChannel();
+    }
 
     const browserGlobals = !process._noBrowserGlobals;
     if (browserGlobals) {
