+++ lib/internal/process/per_thread.js
@@ -1,5 +1,9 @@
 'use strict';
 
+// This files contains process bootstrappers that can be
+// run when setting up each thread, including the main
+// thread and the worker threads.
+
 const {
   errnoException,
   codes: {
@@ -14,19 +18,19 @@ const {
 } = require('internal/errors');
 const util = require('util');
 const constants = process.binding('constants').os.signals;
-const assert = require('assert').strict;
 const { deprecate } = require('internal/util');
-const { isMainThread } = require('internal/worker');
 
-process.assert = deprecate(
-  function(x, msg) {
-    if (!x) throw new ERR_ASSERTION(msg || 'assertion error');
-  },
-  'process.assert() is deprecated. Please use the `assert` module instead.',
-  'DEP0100');
+function setupAssert() {
+  process.assert = deprecate(
+    function(x, msg) {
+      if (!x) throw new ERR_ASSERTION(msg || 'assertion error');
+    },
+    'process.assert() is deprecated. Please use the `assert` module instead.',
+    'DEP0100');
+}
 
 // Set up the process.cpuUsage() function.
-function setup_cpuUsage(_cpuUsage) {
+function setupCpuUsage(_cpuUsage) {
   // Create the argument array that will be passed to the native function.
   const cpuValues = new Float64Array(2);
 
@@ -90,7 +94,7 @@ function setup_cpuUsage(_cpuUsage) {
 // The 3 entries filled in by the original process.hrtime contains
 // the upper/lower 32 bits of the second part of the value,
 // and the remaining nanoseconds of the value.
-function setup_hrtime(_hrtime, _hrtimeBigInt) {
+function setupHrtime(_hrtime, _hrtimeBigInt) {
   const hrValues = new Uint32Array(3);
 
   process.hrtime = function hrtime(time) {
@@ -193,67 +197,6 @@ function setupKillAndExit() {
   };
 }
 
-
-function setupSignalHandlers() {
-  if (!isMainThread) {
-    // Worker threads don't receive signals.
-    return;
-  }
-
-  const signalWraps = Object.create(null);
-  let Signal;
-
-  function isSignal(event) {
-    return typeof event === 'string' && constants[event] !== undefined;
-  }
-
-  // Detect presence of a listener for the special signal types
-  process.on('newListener', function(type) {
-    if (isSignal(type) && signalWraps[type] === undefined) {
-      if (Signal === undefined)
-        Signal = process.binding('signal_wrap').Signal;
-      const wrap = new Signal();
-
-      wrap.unref();
-
-      wrap.onsignal = process.emit.bind(process, type, type);
-
-      const signum = constants[type];
-      const err = wrap.start(signum);
-      if (err) {
-        wrap.close();
-        throw errnoException(err, 'uv_signal_start');
-      }
-
-      signalWraps[type] = wrap;
-    }
-  });
-
-  process.on('removeListener', function(type) {
-    if (signalWraps[type] !== undefined && this.listenerCount(type) === 0) {
-      signalWraps[type].close();
-      delete signalWraps[type];
-    }
-  });
-}
-
-
-function setupChannel() {
-  // If we were spawned with env NODE_CHANNEL_FD then load that up and
-  // start parsing data from that stream.
-  if (process.env.NODE_CHANNEL_FD) {
-    const fd = parseInt(process.env.NODE_CHANNEL_FD, 10);
-    assert(fd >= 0);
-
-    // Make sure it's not accidentally inherited by child processes.
-    delete process.env.NODE_CHANNEL_FD;
-
-    require('child_process')._forkChild(fd);
-    assert(process.send);
-  }
-}
-
-
 function setupRawDebug(_rawDebug) {
   process._rawDebug = function() {
     _rawDebug(util.format.apply(null, arguments));
@@ -288,13 +231,12 @@ function setupUncaughtExceptionCapture(exceptionHandlerState,
 }
 
 module.exports = {
-  setup_cpuUsage,
-  setup_hrtime,
+  setupAssert,
+  setupCpuUsage,
+  setupHrtime,
   setupMemoryUsage,
   setupConfig,
   setupKillAndExit,
-  setupSignalHandlers,
-  setupChannel,
   setupRawDebug,
   setupUncaughtExceptionCapture
 };
