+++ lib/internal/process/stdio.js
@@ -6,21 +6,17 @@ const {
   ERR_UNKNOWN_STDIN_TYPE,
   ERR_UNKNOWN_STREAM_TYPE
 } = require('internal/errors').codes;
-const {
-  isMainThread,
-  workerStdio
-} = require('internal/worker');
 
-exports.setup = setupStdio;
+exports.setupProcessStdio = setupProcessStdio;
+exports.getMainThreadStdio = getMainThreadStdio;
 
-function setupStdio() {
+function getMainThreadStdio() {
   var stdin;
   var stdout;
   var stderr;
 
   function getStdout() {
     if (stdout) return stdout;
-    if (!isMainThread) return workerStdio.stdout;
     stdout = createWritableStdioStream(1);
     stdout.destroySoon = stdout.destroy;
     stdout._destroy = function(er, cb) {
@@ -36,7 +32,6 @@ function setupStdio() {
 
   function getStderr() {
     if (stderr) return stderr;
-    if (!isMainThread) return workerStdio.stderr;
     stderr = createWritableStdioStream(2);
     stderr.destroySoon = stderr.destroy;
     stderr._destroy = function(er, cb) {
@@ -52,8 +47,6 @@ function setupStdio() {
 
   function getStdin() {
     if (stdin) return stdin;
-    if (!isMainThread) return workerStdio.stdin;
-
     const tty_wrap = process.binding('tty_wrap');
     const fd = 0;
 
@@ -136,6 +129,14 @@ function setupStdio() {
     return stdin;
   }
 
+  return {
+    getStdout,
+    getStderr,
+    getStdin
+  };
+}
+
+function setupProcessStdio({ getStdout, getStdin, getStderr }) {
   Object.defineProperty(process, 'stdout', {
     configurable: true,
     enumerable: true,
