+++ node.gyp
@@ -133,12 +133,13 @@
       'lib/internal/net.js',
       'lib/internal/os.js',
       'lib/internal/process/esm_loader.js',
-      'lib/internal/process/methods.js',
+      'lib/internal/process/main_thread_only.js',
       'lib/internal/process/next_tick.js',
+      'lib/internal/process/per_thread.js',
       'lib/internal/process/promises.js',
       'lib/internal/process/stdio.js',
       'lib/internal/process/warning.js',
-      'lib/internal/process.js',
+      'lib/internal/process/worker_thread_only.js',
       'lib/internal/querystring.js',
       'lib/internal/process/write-coverage.js',
       'lib/internal/readline.js',
